requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
WTForms==3.0.1
Werkzeug==2.3.7
reportlab==4.0.4
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
pywhatkit==5.4
matplotlib==3.7.2
plotly==5.17.0
pandas==2.1.1
email-validator==2.0.0
bcrypt==4.0.1
gunicorn==21.2.0
